Add nullable annotations
