+++ Src/ITranslatorClient.cs
@@ -22,14 +22,14 @@ namespace TranslatorService
         /// <remarks>
         /// <para>You must register Microsoft Translator on https://portal.azure.com/#create/Microsoft.CognitiveServicesTextTranslation to obtain the Subscription key needed to use the service.</para>
         /// </remarks>
-        string SubscriptionKey { get; set; }
+        string? SubscriptionKey { get; set; }
 
         /// <summary>
         /// Gets or sets the the Azure region of the the Translator service.
         /// </summary>
         /// <remarks>This value is used to automatically set the <see cref="AuthenticationUri"/> property. If the paramter is <strong>null</strong> (<strong>Nothing</strong> in Visual Basic), the global service is used.
         /// </remarks>
-        string Region { get; set; }
+        string? Region { get; set; }
 
         /// <summary>
         /// Gets or sets the string representing the supported language code to translate the text in.
@@ -60,7 +60,7 @@ namespace TranslatorService
         /// <para>Calling this method isn't mandatory, because the token is get/refreshed everytime is needed. However, it is called at startup, it can speed-up subsequest requests.</para>
         /// <para>You must register Microsoft Translator on https://portal.azure.com/#create/Microsoft.CognitiveServicesTextTranslation to obtain the Subscription key needed to use the service.</para>
         /// </remarks>
-        Task InitializeAsync(string region, string subscriptionKey, string language = null);
+        Task InitializeAsync(string region, string subscriptionKey, string? language = null);
 
         /// <summary>
         /// Detects the language of a text.
@@ -126,7 +126,7 @@ namespace TranslatorService
         /// <para>For more information, go to https://docs.microsoft.com/azure/cognitive-services/translator/reference/v3-0-languages.
         /// </para>
         /// </remarks>
-        Task<IEnumerable<ServiceLanguage>> GetLanguagesAsync(string language = null);
+        Task<IEnumerable<ServiceLanguage>> GetLanguagesAsync(string? language = null);
 
         /// <summary>
         /// Translates a text string into the specified language.
@@ -153,7 +153,7 @@ namespace TranslatorService
         /// </remarks>
         /// <seealso cref="Language"/>
         /// <seealso cref="GetLanguagesAsync"/>
-        Task<TranslationResponse> TranslateAsync(string input, string to = null);
+        Task<TranslationResponse> TranslateAsync(string input, string? to = null);
 
         /// <summary>
         /// Translates a text string into the specified languages.
@@ -311,7 +311,7 @@ namespace TranslatorService
         /// </remarks>
         /// <seealso cref="Language"/>
         /// <seealso cref="GetLanguagesAsync"/>
-        Task<IEnumerable<TranslationResponse>> TranslateAsync(IEnumerable<string> input, IEnumerable<string> to = null);
+        Task<IEnumerable<TranslationResponse>> TranslateAsync(IEnumerable<string> input, IEnumerable<string>? to = null);
 
         /// <summary>
         /// Translates a list of sentences into the specified languages.
