+++ Src/Models/Speech/RecognitionAlternative.cs
@@ -18,23 +18,23 @@ namespace TranslatorService.Models.Speech
         /// The lexical form of the recognized text: the actual words recognized.
         /// </summary>
         [JsonPropertyName("Lexical")]
-        public string LexicalForm { get; set; }
+        public string LexicalForm { get; set; } = null!;
 
         /// <summary>
         /// The inverse-text-normalized ("canonical") form of the recognized text, with phone numbers, numbers, abbreviations ("doctor smith" to "dr smith"), and other transformations applied.
         /// </summary>
         [JsonPropertyName("ITN")]
-        public string CanonicalForm { get; set; }
+        public string CanonicalForm { get; set; } = null!;
 
         /// <summary>
         /// The ITN form with profanity masking applied, if requested.
         /// </summary>
         [JsonPropertyName("MaskedITN")]
-        public string MaskedCanonicalForm { get; set; }
+        public string MaskedCanonicalForm { get; set; } = null!;
 
         /// <summary>
         /// The display form of the recognized text, with punctuation and capitalization added. This parameter is the same as <seealso cref="SpeechRecognitionResponse.DisplayText"/> provided when <seealso cref="RecognitionResultFormat"/> is set to <seealso cref="RecognitionResultFormat.Simple"/>.
         /// </summary>
-        public string Display { get; set; }
+        public string Display { get; set; } = null!;
     }
 }
