+++ Src/Models/Translation/Translation.cs
@@ -9,11 +9,11 @@
         /// <summary>
         /// Gets or sets a string giving the translated text.
         /// </summary>
-        public string Text { get; set; }
+        public string Text { get; set; } = null!;
 
         /// <summary>
         /// Gets or sets a string representing the language code of the target language.
         /// </summary>
-        public string To { get; set; }
+        public string To { get; set; } = null!;
     }
 }
